Fix issue with DropPartitionsRequest and defaults
When using the DropPartitionsRequest call, the IfExists and NeedResult
fields both have a default of "true", unfortunately with how the thrift
compiler generates the go code it doesn't send optional fields with
defaults even if you explicitly set them. Hive doesn't interpret this
the same way and so there's no way with the generated code to ever send
"true" for either of those fields.